added rat anims
