+++ LazyNecromancer/Assets/Scenes/GREG.unity
@@ -132,6 +132,9 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 110014762}
+  - component: {fileID: 110014765}
+  - component: {fileID: 110014763}
+  - component: {fileID: 110014764}
   m_Layer: 0
   m_Name: POC2
   m_TagString: Untagged
@@ -147,12 +150,70 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 110014761}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 1.5600698, y: 1.5495251, z: -11.301469}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &110014763
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 110014761}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 857f8e3306361024eb8c51d5ff63eeb3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  settings: {fileID: 11400000, guid: 7e2f41643edde9940991a40ca22df23d, type: 2}
+  grid: {fileID: 110014764}
+  target: {fileID: 110014762}
+  spawnCount: 25
+  numBoids: 100
+  maxFlockSize: 100
+  DEBUG: 0
+  spawnRadius: 10
+  prefab: {fileID: 5092761694481593269, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e, type: 3}
+--- !u!114 &110014764
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 110014761}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e5c8bafc8cd619c43bdab1de7d163372, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  numOfFilled: 0
+  boidSettings: {fileID: 11400000, guid: 7e2f41643edde9940991a40ca22df23d, type: 2}
+--- !u!114 &110014765
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 110014761}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c8ced8fa28fec904bbfcd558d43f0602, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  spawnRadius: 10
+  spawnCount: 426
+  fireRate: 2
+  spawnBoids: 1
+  sprites: []
+  colors: []
+  currentColor: {r: 1, g: 1, b: 1, a: 1}
+  gridOrigin: {x: -250, y: 250, z: 0}
+  size: 500
+  cellSize: 1
+  prefab: {fileID: 7799068297280379563, guid: 6b4ad74547957154c843b013932debab, type: 3}
 --- !u!1 &402382387
 GameObject:
   m_ObjectHideFlags: 0
@@ -465,7 +526,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!114 &1073135533
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -520,7 +581,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &1290272120
 RectTransform:
   m_ObjectHideFlags: 0
@@ -609,7 +670,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &1475656180
 RectTransform:
   m_ObjectHideFlags: 0
@@ -875,3 +936,77 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
+--- !u!1001 &5092761693320720963
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 5092761694481593266, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_Name
+      value: RatPrefab
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593266, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 5
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5092761694481593271, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 5e15369f1e18f8a4b9cf51fa3acbe73e, type: 3}
