+++ LazyNecromancer/Assets/Scripts/Rats/EnemyBoidController.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyBoidController : MonoBehaviour
+{
+
+    
+    void Start() {
+        
+    }
+
+    void Update() {
+        
+    }
+}
