+++ LazyNecromancer/Assets/Scripts/Rats/EnemyBoidController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1c4f7ca0f3a75144690206e727fe92bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
