+++ LazyNecromancer/Assets/Scripts/Rats/RatAnimController.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3e70d4f45993938478dd96e7eddaa475
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
