+++ LazyNecromancer/Assets/Scripts/Rats/RatBoidSettings.asset
@@ -1,0 +1,29 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6b081e9043a50fe4bacef5755f4ac89e, type: 3}
+  m_Name: RatBoidSettings
+  m_EditorClassIdentifier: 
+  minSpeed: 10
+  maxSpeed: 50
+  perceptionRadius: 2.5
+  avoidanceRadius: 1
+  maxSteerForce: 3
+  alignWeight: 1
+  cohesionWeight: 1
+  seperateWeight: 1.5
+  targetWeight: 0.1
+  obstacleMask:
+    serializedVersion: 2
+    m_Bits: 0
+  boundsRadius: 0.3
+  avoidCollisionWeight: 30
+  collisionAvoidDst: 5
