+++ LazyNecromancer/Assets/Scripts/Rats/RatBoidSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e2f41643edde9940991a40ca22df23d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
