+++ LazyNecromancer/Assets/Scripts/Rats/RatDeath.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b87214c4b44f3fc40852d3c77b1326b6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
