+++ LazyNecromancer/Assets/Scripts/Rats/RatMoveAnim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7207e250d2bfe944a5f614593176d3c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
