+++ LazyNecromancer/Assets/Scripts/Rats/Spawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8ced8fa28fec904bbfcd558d43f0602
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
